Utils: Move offer_missing_install to libs/installation.py
